% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{combine_pdF}
\alias{combine_pdF}
\title{Combine multiple delta-F p values using Fisher method}
\usage{
combine_pdF(x, df_format = FALSE, varname = "pdF")
}
\arguments{
\item{x}{a data frame with at least one variable containing p-values from \code{deltaF.linear}}

\item{df_format}{Should results be returned as a data frame?}

\item{varname}{name of variable containing p-values}
}
\value{
either a single p-value or data.frame containing p-value
}
\description{
Combine multiple delta-F p values using Fisher method
}
\details{
Just a simple wrapper of \code{Fisher.method}
}
\seealso{
\code{\link{Fisher.method}}
}

