% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{combine_V}
\alias{combine_V}
\title{Combine multiple V statistics by taking the product}
\usage{
combine_V(x, df_format = FALSE, varname = "V")
}
\arguments{
\item{x}{a data frame with at least one variable containing V statistics from \code{deltaF.linear}}

\item{df_format}{Should results be returned as a data frame?}

\item{varname}{name of variable containing V statistics}
}
\value{
either a single V stat or data.frame containing V stat
}
\description{
Combine multiple V statistics by taking the product
}
\details{
Just a simple wrapper of \code{evi.valMult}
}
\seealso{
\code{\link{evi.valMult}}
}

