% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_veracity2.R
\name{deltaF.linear}
\alias{deltaF.linear}
\title{Function that computes F-test for nested (ANOVA) models}
\usage{
deltaF.linear(meanVec, sdVec, n.fl)
}
\arguments{
\item{meanVec}{numeric vector containing means for each factor level}

\item{sdVec}{numeric vector containing standard deviations for each factor level}

\item{n.fl}{(numeric) integer indicating cell size/nr. of replicates per factor level}
}
\value{
A list with delta F and associated p-value.
}
\description{
Function that computes F-test for nested (ANOVA) models
}
\details{
\itemize{
	\item Assumes only summary/descriptive statistics are available (from publication under investigation)
	\item Assumes ANOVA-type setting with one-way factorial design
	\item Assumes the ANOVA is balanced (nr. of replicates same for each factor level)
	\item Assumes the input vectors are ordered according to linearity of effect, e.g.,
	  meanVec[1] < meanVec[2] < ... < meanVec[k]
	\item Compares reduced linear model with more complex model bearing k - 1 regression parameters
	\item Null hypothesis: perfect linearity in underlying means
	\item Function based on whistleblower report in JF case
 \item Unmodified function from UvA report by Peeters, Klaasen, and van de Wiel (2015)
}
}
\references{
Peeters, C. F. W., Klaasen, C. A. J., and van de Wiel, M. A. (2015).  \emph{Evaluating the Scientific Veracity of Publications by dr. Jens Förster.}  Retrieved from \href{http://retractionwatch.files.wordpress.com/2014/04/report_foerster.pdf}{Retraction Watch website}.
}

