% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{deltaFp}
\alias{deltaFp}
\title{Wrapper function for calculating delta F}
\usage{
deltaFp(x, df_format = FALSE)
}
\arguments{
\item{x}{A numeric vector (or row from a data frame) containing means, standard deviations, and sample size per group as named elements \code{m1}, \code{m2}, \code{m3} and \code{s1}, \code{s2}, and \code{s3}, and \code{N}, respectively.}

\item{df_format}{Should results be returned as a data frame? (Useful when calling the function from \code{\link[dplyr]{do}}).}
}
\value{
p-value for delta F (single numeric value)
}
\description{
Wrapper function for calculating delta F
}
\details{
This function wraps \code{\link{deltaF.linear}}, returning the associated p-value.
}
\seealso{
\code{\link{deltaF.linear}}
}

