% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{discretize}
\alias{discretise}
\alias{discretize}
\title{Convert continuous to interval data}
\usage{
discretize(x, boundaries = seq(-3, 3, length.out = 4))
}
\arguments{
\item{x}{the continuous data}

\item{boundaries}{boundaries defining how continuous variable is
mapped to scale levels}
}
\value{
vector with discrete ratings
}
\description{
Convert continuous to interval data
}
\details{
Splits up a continuous scale into a discrete scale.  The
scale will have a number of levels equal to one more than the
number of elements in \code{boundaries}.  The default behavior is
to create an interval scale with upper and lower boundaries at -3
and 3 respectively.  This function is a fairly simple wrapper for
\code{\link{cut}}.
}
\seealso{
\code{link{cut}}
}

