% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_veracity2.R
\name{evi.val}
\alias{evi.val}
\title{Function that computes an evidential value in favor of dependent observations versus independence}
\usage{
evi.val(meanVec, sdVec, n.fl)
}
\arguments{
\item{meanVec}{numeric vector containing means for each factor level}

\item{sdVec}{numeric vector containing standard deviations for each factor level}

\item{n.fl}{(numeric) integer indicating cell size/nr. of replicates per factor level}
}
\value{
single-element vector with V statistic, or a list with upper and lower limits for V
}
\description{
Function that computes an evidential value in favor of dependent observations versus independence
}
\details{
Function was included without modification from UvA report by Peeters, Klaasen,
and van de Wiel (2015).
\itemize{
\item Assumes only summary/descriptive statistics are available (from publication under investigation)
\item Assumes ANOVA-type setting with one-way factorial design
\item Assumes there are 3 factor levels
\item Assumes the ANOVA is balanced (nr. of replicates same for each factor level)
\item Assumes following restriction on group means: mu1 \item 2*mu2 + mu3 = 0
\item Assumes the input vectors are ordered according to linearity of effect, e.g., meanVec[1] < meanVec[2]  < meanVec[3]
\item Assumes dependence between the measurement errors undermines (the assumption of) veracity
\item Computes the evidential value in favor of the hypothesis of a dependence-structure in the underlying
data (i.e., correlated measurement errors, indicating incorrect data collection: H_f) versus the
hypothesis of independence (the standard ANOVA assumption: H_i)
\item The evidential value as calculated is defined in Klaassen (2015).
}
}
\references{
Peeters, C. F. W., Klaasen, C. A. J., and van de Wiel, M. A. (2015).  \emph{Evaluating the Scientific Veracity of Publications by dr. Jens Förster.}  Retrieved from \href{http://retractionwatch.files.wordpress.com/2014/04/report_foerster.pdf}{Retraction Watch website}.

Klaassen, C.A.J. (2015). \emph{Evidential Value in ANOVA-Regression Results in Scientific Integrity Studies}. \href{http://arxiv.org/pdf/1405.4540v2}{Arxiv: 1405.4540v2}.
}

