% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_veracity2.R
\name{evi.valMult}
\alias{evi.valMult}
\title{Function that computes overall evidential value}
\usage{
evi.valMult(EVvals)
}
\arguments{
\item{EVvals}{a vector containing multiple evidential values (ideally stemming from 'evi.val')}
}
\value{
Product of all elements in \code{EVvals}
}
\description{
Function that computes overall evidential value
}
\details{
Computes overall evidential value by simple product of
evidential values (of individual studies).  When, from 'evi.val' we
have that SL >= nz2, the evidential value has a range, it is
recommended to use the minimum value for computation of the overall
EV.

Function was included without modification from UvA report by Peeters, Klaasen,
and van de Wiel (2015).
}
\references{
Peeters, C. F. W., Klaasen, C. A. J., and van de Wiel, M. A. (2015).  \emph{Evaluating the Scientific Veracity of Publications by dr. Jens Förster.}  Retrieved from \href{http://retractionwatch.files.wordpress.com/2014/04/report_foerster.pdf}{Retraction Watch website}.

Klaassen, C.A.J. (2015). \emph{Evidential Value in ANOVA-Regression Results in Scientific Integrity Studies}. \href{http://arxiv.org/pdf/1405.4540v2}{Arxiv: 1405.4540v2}.
}

