% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{eviVal}
\alias{eviVal}
\title{Wrapper function for calculating Klaassen's V (evidential value)}
\usage{
eviVal(x, df_format = FALSE)
}
\arguments{
\item{x}{A numeric vector (or row from a data frame) containing means, standard deviations, and sample size per group as named elements \code{m1}, \code{m2}, \code{m3} and \code{s1}, \code{s2}, and \code{s3}, and \code{N}, respectively.}

\item{df_format}{Should results be returned as a data frame? (Useful when calling the function from \code{\link[dplyr]{do}}).}
}
\value{
V, or the lower bound of the range for V
}
\description{
Wrapper function for calculating Klaassen's V (evidential value)
}
\details{
This function wraps \code{\link{evi.val}}, returning the associated V statistic or lower bound if a range is returned.
}
\seealso{
\code{\link{evi.val}}
}

