% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{findings}
\alias{findings}
\title{Findings for studies investigated in UvA Förster report}
\format{\preformatted{'data.frame':	209 obs. of  6 variables:
 $ pub  : chr  "Hagtvedt" "Hagtvedt" "Hunt" "Jia" ...
 $ study: chr  "e1" "e2" "e1" "e1" ...
 $ dF   : num  0.285 0.348 1.515 1.044 0.932 ...
 $ p(dF): num  0.595 0.557 0.222 0.309 0.336 ...
 $ EV   : num  1.4 1.17 1 1 1 ...
 $ EV.up: num  NA NA NA NA NA ...
}}
\usage{
findings
}
\description{
A dataset with the findings for each of the studies in the report
}
\details{
\describe{
\item{pub}{Unique identifying code for the publication}
\item{study}{Identifying code for the study (unique within the publication)}
\item{dF}{Delta F}
\item{p(dF)}{p-value from the delta-F test}
\item{EV}{Evidential Value}
\item{EV.up}{Evidential Value upper limit}
}
}
\keyword{datasets}

