% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_veracity2.R
\name{Fisher.method}
\alias{Fisher.method}
\title{Function that performs Fisher's combined probability test}
\usage{
Fisher.method(pvalVec, plot = FALSE, x.label = "p-values", main = "")
}
\arguments{
\item{pvalVec}{a vector containing p-values (ideally stemming from 'deltaF.linear')}

\item{plot}{boolean specifying whether to produce a plot}

\item{x.label}{character specifying the label for the x-axis, default = "p-values"; only when plot = TRUE}

\item{main}{character specifying the main title, default = ""; only when plot = TRUE}
}
\value{
single scalar value
}
\description{
Function that performs Fisher's combined probability test
}
\details{
Only use to combine results from independent tests that
regard the same (null) hypothesis. The blue line in the plot
represents expectation under uniform behavior.  Function was
included without modification from UvA report by Peeters, Klaasen,
and van de Wiel (2015).
}
\references{
Peeters, C. F. W., Klaasen, C. A. J., and van de Wiel, M. A. (2015).  \emph{Evaluating the Scientific Veracity of Publications by dr. Jens Förster.}  Retrieved from \href{http://retractionwatch.files.wordpress.com/2014/04/report_foerster.pdf}{Retraction Watch website}.
}

