% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{pubs}
\alias{pubs}
\title{List of publications investigated in UvA Förster report}
\format{A data frame with 24 rows and 2 columns: \describe{
\item{pub}{Unique identifying code for the publication}
\item{section}{Classifcation of the publication in the report}
}}
\usage{
pubs
}
\description{
A dataset listing the publications and their classification.
}
\keyword{datasets}

