% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{studies}
\alias{studies}
\title{List of studies investigated in UvA Förster report}
\format{A data frame with 209 rows and 3 columns:
\describe{
\item{pub}{Unique identifying code for the publication}
\item{study}{Identifying code for the study (unique within the publication)}
\item{N}{Total number of observations}
}}
\usage{
studies
}
\description{
A dataset listing the studies (experiments) within the publications
}
\keyword{datasets}

