% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{study_stats}
\alias{study_stats}
\title{Statistics of studies investigated in UvA Förster report}
\format{A data frame with 209 rows and 8 columns:
\describe{
\item{pub}{Unique identifying code for the publication}
\item{study}{Identifying code for the study (unique within the publication)}
\item{m1}{Mean for group 1 (lowest mean)}
\item{m2}{Mean for group 2 (middle mean)}
\item{m3}{Mean for group 3 (highest mean)}
\item{s1}{Standard deviation for group 1}
\item{s2}{Standard deviation for group 2}
\item{s3}{Standard deviation for group 3}
}}
\usage{
study_stats
}
\description{
A dataset with the statistics for the studies investigated
}
\keyword{datasets}

